% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hola.R
\name{hola}
\alias{hola}
\title{Saludos}
\usage{
hola(x)
}
\arguments{
\item{x}{Un \code{character(1)} con tu nombre.}
}
\value{
Imprime un mensaje con el saludo.
}
\description{
Esta función te saluda desde ConectaR2021
}
\examples{

hola("Olga")
}
